requirements.txt
fastapi
uvicorn[standard]
streamlit
google-genai
faster-whisper
python-dotenv
python-multipart
pydantic
fpdf2
notion-client
pandas